build: automated deployment
Configures `semantic-release`'s APM configuration to automatically 
publish a release whenever a commit is made to `master` that should 
require a release, according to the Conventional Commits standard.

Also adds a commit-msg hook automatically installed when `npm install` 
is ran that checks all commit messages, and a Travis CI build step to 
verify that the commit messages in any PRs follow this format.